requirements.txt
django
django-model-utils